requirements.txt
anyio==4.9.0
APScheduler==3.9.1.post1
blinker==1.9.0
cachetools==5.3.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.13.0
python-dotenv==1.1.0
python-telegram-bot==20.0
pytz==2025.2
requests==2.32.3
rfc3986==1.5.0
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
